Trebuchet: keep dock QSB disabled by default
We have two different builds and many users who don't have google app installed,
in that case if we keep it enabled it will create a weird space in the hotseat.
So, keep it off by default for all users.

While we are disabling it why not disable the weird looking all apps arrow indicator.
Because of our hotseat changes the arrow looks very weird and pixelated So, let's keep it disabled under all cases.

Signed-off-by: LordShenron <shen.priyanshu@gmail.com>